Extract score structs into separate module
